Allocate more memory to lcropped
Doesn't appear to fix the runtime issues that are discussed in issue #2, but it seems to be a good idea, as if the left crop value is zero, then I might not have enough room for NULL otherwise.

The previous code assumed that at least one character was being cropped, and thus, thought it was safe to allocate lcrop_len number of characters to the character array.