Prevent over-reading of initial_str
Introduces an if statement that catches when the end of initial_str has be reached. It is part of an attempt to fix running the program on Windows, as detailed in issue #2.

Doesn't appear to fix the problem, but seems helpful to have, as it will prevent some unnecessary looping.