Fix snprintf memory leak
Before this change, valgrind on Linux was always complaining about
invalid writes and reads. This change hopefully fixes it by adding a new
variable for use with snprintf. Before, an incremented f_elements was
used with snprintf, but now, it uses the f_total variable, which will
lead to more predictable results.